Multipath Testing E2E
This test included three VMs E96 with an attached shared ultra disk.
Two VMs served the target role and were running the SPDK stack with a
shared ultra disk attached.

The other VM served as the initiator and is where the fio benchmark was
run.  The controllers and subsystems were attached using nvme cli tool.